Updated PKCE.js to only handle pkce code challenge generation. Moved all auth logic into oauth.js
